% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspect_ratio_FB.R
\name{aspect_ratio_FB}
\alias{aspect_ratio_FB}
\title{Obtain aspect ratio of caudal fin from FishBase}
\usage{
aspect_ratio_FB(
  species_list,
  check_names = FALSE,
  mc.cores = getOption("mc.cores", 1L)
)
}
\arguments{
\item{species_list}{A vector of taxonomic names in the form "Genus species"}

\item{check_names}{Logical, check whether species names are correct
according to the accepted taxonomy in FishBase}

\item{mc.cores}{The number of cores to use for \code{\link[parallel:mclapply]{parallel::mclapply()}}.
Must be exactly 1 on Windows (which uses the master process)}
}
\value{
A tibble with species name, aspect ratio, and taxonomic level.
}
\description{
Obtain average species aspect ratio
from FishBase at the lowest available taxonomic level,
either species, genus or family level.
}
\examples{
\dontrun{
aspect_ratio_FB("Caranx ignobilis")
aspect_ratio_FB(c("Chaetodon auriga", "Caesio cuning"))
library(rfishbase)
scombrids <- species_list(Family = "Scombridae")
aspect_ratio_FB(scombrids)
}

}
