% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass_FB.R
\name{biomass_FB}
\alias{biomass_FB}
\title{Obtain fish biomass}
\usage{
biomass_FB(data, check_names = FALSE)
}
\arguments{
\item{data}{A data frame with two columns:
\itemize{
\item{"species":} taxonomic names in the form "Genus species"
\item{"length":} total length in "cm"
}}

\item{check_names}{Logical, check whether species names are correct
according to the accepted taxonomy in FishBase}
}
\value{
The input data frame with an additional column for biomass


}
\description{
Compute fish biomass using a and b parameters from FishBase knowing total length
}
\examples{
\dontrun{
biomass_FB(data.frame(species = c("Dentex dentex", "Chaetodon lunula"),
                      length = c(20, 12)),
           check_names = TRUE)
}

}
