% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass_FB.R
\name{biomass_FB_sp}
\alias{biomass_FB_sp}
\title{Obtain fish biomass for taxa identified to genus level}
\usage{
biomass_FB_sp(data)
}
\arguments{
\item{data}{A data frame with two columns:
\itemize{
\item{"species":} taxonomic names in the form "Genus sp." or "Genus spp."
\item{"length":} total length in "cm"
}}
}
\value{
The input data frame with an additional column for biomass


}
\description{
Compute fish biomass using genus-average a and b parameters from FishBase knowing total length
}
\examples{
\dontrun{
biomass_FB_sp(data.frame(species = c("Dentex sp.", "Chaetodon spp."),
                         length = c(20, 12)))
}

}
