% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_name_fishtree.R
\name{check_name_fishtree}
\alias{check_name_fishtree}
\title{Check species names and sampled species in the Fish Tree of Life}
\usage{
check_name_fishtree(species_list, sampled = FALSE)
}
\arguments{
\item{species_list}{A vector of taxonomic names in the form "Genus species"}

\item{sampled}{Logical, if the list has to be checked for unsampled species
(i.e. those without genetic data, see Rabosky et al. 2018 for detailed methodology)}
}
\value{
If "sampled=FALSE", a message and a vector of wrong species names (if any).
If "sampled=TRUE", a message and a list with two components:
\itemize{
\item{"name_error":} wrong species names
\item{"not_sampled":} species without genetic data
}
}
\description{
Check species names and sampled species in the Fish Tree of Life
}
\examples{
\dontrun{
check_name_fishtree("Boops boops")
check_name_fishtree(c("Balistapus undulatus", "Plectropomus leopardus"),  sampled = TRUE)
}

}
\references{
Rabosky D. L. et al. (2018) An inverse latitudinal gradient in
speciation rate for marine fishes. Nature, 559, 392–395.
https://doi.org/10.1038/s41586-018-0273-1
}
\author{
Mattia Ghilardi, \email{mattia.ghilardi91@gmail.com}
}
