% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_names_FB.R
\name{check_names_FB}
\alias{check_names_FB}
\title{A modified version of "fishflux::check_name_fishbase()" that returns wrong names}
\usage{
check_names_FB(species_list)
}
\arguments{
\item{species_list}{A vector of taxonomic names in the form "Genus species"}
}
\description{
A modified version of "fishflux::check_name_fishbase()" that returns wrong names
}
