% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elongation_FB.R
\name{elongation_FB}
\alias{elongation_FB}
\title{Obtain body elongation from FishBase}
\usage{
elongation_FB(
  species_list,
  check_names = FALSE,
  mc.cores = getOption("mc.cores", 1L)
)
}
\arguments{
\item{species_list}{A vector of taxonomic names in the form "Genus species"}

\item{check_names}{Logical, check whether species names are correct
according to the accepted taxonomy in FishBase}

\item{mc.cores}{The number of cores to use for \code{\link[parallel:mclapply]{parallel::mclapply()}}.
Must be exactly 1 on Windows (which uses the master process)}
}
\value{
A tibble with species name, body elongation, and taxonomic level.
}
\description{
Obtain average species elongation (i.e., standard length/maximum body depth)
from FishBase at the lowest available taxonomic level,
either species, genus or family level.
}
\examples{
\dontrun{
elongation_FB("Sarpa salpa")
elongation_FB(c("Aulostomus chinensis", "Chaetodon lunulatus"))
library(rfishbase)
scombrids <- species_list(Family = "Scombridae")
elongation_FB(scombrids)
}

}
