% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elongation_FB.R
\name{elongation_FB_sp}
\alias{elongation_FB_sp}
\title{Obtain body elongation from FishBase for taxa identified to genus level}
\usage{
elongation_FB_sp(species_list, mc.cores = getOption("mc.cores", 1L))
}
\arguments{
\item{species_list}{A vector of taxonomic names in the form "Genus sp." or "Genus spp."}

\item{mc.cores}{The number of cores to use for \code{\link[parallel:mclapply]{parallel::mclapply()}}.
Must be exactly 1 on Windows (which uses the master process)}
}
\value{
A tibble with the provided name, body elongation, and taxonomic level.
}
\description{
Obtain average species elongation (i.e., standard length/maximum body depth)
from FishBase at the lowest available taxonomic level,
either genus or family level.
}
\examples{
\dontrun{
elongation_FB_sp("Sarpa sp.")
elongation_FB_sp(c("Epinephelus sp.", "Sillago spp."))
}

}
