% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fix_ran.R
\name{extract_fix_ran}
\alias{extract_fix_ran}
\title{Extract names of fixed and random effects from brms models}
\usage{
extract_fix_ran(model)
}
\arguments{
\item{model}{brmsfit object}
}
\description{
Extract names of fixed and random effects from brms models
}
