% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_traits.R
\name{family_traits}
\alias{family_traits}
\title{Obtain family-average trophic level, caudal fin aspect ratio, body elongation, and relative intestinal length}
\usage{
family_traits(
  family_list,
  type = c("both", "diet", "food items"),
  mc.cores = getOption("mc.cores", 1L),
  ...
)
}
\arguments{
\item{family_list}{A vector of taxonomic families}

\item{type}{Character, indicating the type of trophic level desired. Possible values are:
"both" (default), "diet", "food items" or any abbreviation of these (see Details)}

\item{mc.cores}{The number of cores to use for \code{\link[parallel:mclapply]{parallel::mclapply()}}.
Must be exactly 1 on Windows (which uses the master process)}

\item{...}{Additional arguments passed to intestinal_length()}
}
\value{
A tibble with species name, trophic level, caudal fin aspect ratio,
body elongation and relative intestinal length.
}
\description{
Obtain family-average trophic level, caudal fin aspect ratio, body elongation, and relative intestinal length
}
\details{
The function can retrieve three different trophic level estimates:
\itemize{
\item{\code{type="both"}:} computes the average of the trophic levels based on
diet and food items when both are available,
otherwise return the available estimate;
\item{\code{type="diet"}:} retrieves trophic level based on diet composition only.
These estimates are available only for some species.
Note that for some families there are no diet studies available;
\item{\code{type="food items"}:} retrieves trophic level based on food items only.
}
}
