% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intestinal_length_sp.R
\name{intestinal_length_sp}
\alias{intestinal_length_sp}
\title{Estimate intestinal length for fishes identified to genus level}
\usage{
intestinal_length_sp(
  data,
  method = "fitted",
  ndraws = 1000,
  seed = 1,
  mc.cores = getOption("mc.cores", 1L),
  summary = TRUE,
  .point = mean,
  .width = 0.95
)
}
\arguments{
\item{data}{A data frame with five columns:
\itemize{
\item{"family":} the taxonomic family
\item{"genus":} the taxonomic genus
\item{"species":} the taxonomic name in the form "Genus sp." or "Genus spp."
\item{"sl":} standard length in "mm"
\item{"elongation":} body elongation (i.e. standard length/max body depth)
\item{"trophic_level":} trophic level (from 1 to 5, preferably based on food items)
}}

\item{method}{Either 'fitted' or 'predict'. Predictions performed with 'predict'
have higher variance as they incorporate the residual error.
However, the estimated means of both methods should be very similar}

\item{ndraws}{The number of draws to be used for the predictions (defaults to 1000)}

\item{seed}{The seed for random number generation to make results reproducible.
Must be different from NULL}

\item{mc.cores}{The number of cores to use for
\code{\link[parallel:mclapply]{parallel::mclapply()}}.
Must be exactly 1 on Windows (which uses the master process)}

\item{summary}{Logical, if predictions should be summarised.
If TRUE returns mean and sd, if FALSE returns all draws}

\item{.point}{Only if 'summary=TRUE', point summary function,
e.g. 'mean' or 'median', default to 'mean'}

\item{.width}{Only if 'summary=TRUE', according to
\code{\link[tidybayes:point_interval]{tidybayes::point_interval()}}}
}
\value{
A data frame with natural-log intestinal length in "mm"
}
\description{
Predict genus-level intestinal length of fish based on phylogeny, body morphology, and trophic level
}
\examples{
\dontrun{
dummy <- data.frame(family = c("Siganidae", "Sparidae"),
                    genus = c("Siganus", "Sparus"),
                    species = c("Siganus sp.", "Sparus sp."),
                    sl = c(120, 160))
d_elon <- elongation_FB_sp(dummy$species)
d_troph <- trophic_level_FB_sp(dummy$species, type = "food items")
dummy <- left_join(dummy, d_elon[, -3]) \%>\%
  left_join(d_troph[, -3])
intestinal_length_sp(dummy, ndraws = 10)
intestinal_length_sp(dummy, ndraws = 10, summary = FALSE)
intestinal_length_sp(dummy, ndraws = 10, .point = median, .width = c(0.5, 0.95))
}

}
