% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_variation.R
\name{percent_variation}
\alias{percent_variation}
\title{Calculate \% of variation in the response between two points along the range of a predictor "x"}
\usage{
percent_variation(
  model,
  x1,
  x2,
  exp = TRUE,
  nsamples = NULL,
  summary = TRUE,
  robust = FALSE,
  probs = c(0.025, 0.975)
)
}
\arguments{
\item{model}{A "brmsfit" object}

\item{x1}{Data frame with a value of "x" and the mean value for the other continuous predictors (if any)}

\item{x2}{Same as x1 but with a higher value of "x"}

\item{exp}{Logical, to back-transform when the response variable is modelled on the natural-log scale; default to TRUE}

\item{nsamples}{Positive integer indicating how many posterior samples should be used.
If NULL (the default) all samples are used.}

\item{summary}{Logical, if TRUE (the default) summary statistics are returned,
if FALSE returns raw values}

\item{robust}{Only if "summary=TRUE". If FALSE (the default) the central tendency estimate is the mean,
if TRUE is the median}

\item{probs}{Only if "summary=TRUE". The percentiles to be computed by the quantile function}
}
\value{
A value
}
\description{
It computes the average \% of variation in the response variable of a brms regression
model by the predictor "x"
}
