% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_effect.R
\name{phylo_effect}
\alias{phylo_effect}
\title{Estimate phylogenetic effect for unobserved taxa from a Bayesian phylogenetic model}
\usage{
phylo_effect(
  model,
  ...,
  ndraws = NULL,
  seed = NULL,
  mc.cores = getOption("mc.cores", 1L),
  phy
)
}
\arguments{
\item{model}{A Bayesian phylogenetic model of class 'brmsfit'}

\item{...}{Expression in the form of \code{variable_name[dimension_1, ]}. (see \code{\link[tidybayes:spread_draws]{tidybayes::spread_draws()}})}

\item{ndraws}{The number of draws to return, or NULL to return all draws}

\item{seed}{A seed to use when subsampling draws (i.e. when ndraw is not NULL)}

\item{mc.cores}{The number of cores to use for \code{\link[parallel:mclapply]{parallel::mclapply()}}.
Must be exactly 1 on Windows (which uses the master process)}

\item{phy}{An object of class "phylo" or "multiPhylo" including the new taxa
for which the phylogenetic effect has to be predicted}
}
\value{
A data frame
}
\description{
\code{phylo_effect()} uses \code{\link[picante:phyEstimate]{picante::phyEstimate()}}
to estimate the phylogenetic effect for unobserved taxa through phylogenetic
ancestral state reconstruction. The estimation is repeated \code{ndraws} times to provide
a distribution comparable to that provided by the model for observed taxa.
The function is available for one term only (i.e. the intercept).
}
\examples{
\dontrun{
# Imagine you have a data frame called 'data' with these columns:
# 'y' - the response variable
# 'x1' - predictor 1
# 'x2' - predictor 2
# 'phylo' - taxonomic names
# Then you have a phylogenetic tree (class "phylo" or "multiphylo") for these species called 'tree1'
# You can create a phylogenetic correlation matrix using ape::vcv.phylo(..., corr = TRUE)
A <- ape::vcv.phylo(tree1, corr = TRUE)
# Then fit a phylogenetic model in brms
m1 <- brm(y ~ x1 + x2 + (1|gr(phylo, cov = A)),
          data = data, family = gaussian(), data2 = list(A = A), ...)
# With this model you can make predictions for the observed taxa
# (i.e., those used to train the model), but if you would like to make
# predictions for unobserved taxa you can only use the parameter 'allow_new_levels=TRUE'.
# However, this includes the uncertainty associated with the random intercept
# in the prediction, but does not account for the phylogenetic structure.
# To account for the phylogeny we could predict the phylogenetic effect for
# the unobserved species using 'phylo_effect()'.
# You need a second phylogenetic tree called 'tree2', which is 'tree1'
# with the additional taxa for which you would like to predict y and for which
# x1 and x2 are available.
phylo_eff_pred <- phylo_effect(model = m1, r_phylo[species,],
                               ndraws = 1000, seed = 123, phy = tree2)
Now "y" can be predicted for these new species
}

}
