% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_obs_vs_pred.R
\name{plot_obs_vs_pred}
\alias{plot_obs_vs_pred}
\title{Plot observed vs predicted values}
\usage{
plot_obs_vs_pred(
  data = NULL,
  obs,
  pred,
  xlab,
  ylab,
  smooth_size = 0.7,
  smooth_alpha = 0.3,
  smooth_color = "#0D0887FF",
  smooth_fill = "#0D0887FF",
  point_size = 0.8,
  point_alpha = 0.5,
  point_stroke = 0.1,
  point_shape = 19,
  point_color = "#0D0887FF",
  point_fill = "#0D0887FF",
  text_size = 3,
  abline_size = 0.7,
  abline_color = "red",
  abline_type = 2
)
}
\arguments{
\item{data}{A data frame containing observed and predicted values}

\item{obs}{A numeric vector, or character string if \code{data} is supplied, of observed values}

\item{pred}{A numeric vector, or character string if \code{data} is supplied, of predicted values}

\item{xlab}{A label for the x axis (predicted values)}

\item{ylab}{A label for the y axis (observed values)}

\item{smooth_size}{The size of the regression line}

\item{smooth_alpha}{The opacity of the confidence interval around the regression line}

\item{smooth_color}{The color of the regression line}

\item{smooth_fill}{The color of the confidence interval around the regression line}

\item{point_size}{The size of the points}

\item{point_alpha}{The opacity of the points}

\item{point_stroke}{The width of the points' border}

\item{point_shape}{The shape of the points}

\item{point_color}{The color of the points}

\item{point_fill}{The color of the inside of shapes with a border}

\item{text_size}{The size of the equation and R2}

\item{abline_size}{The size of the reference line}

\item{abline_color}{The color of the reference line}

\item{abline_type}{The type of reference line}
}
\value{
A ggplot object
}
\description{
Display a scatterplot and linear regression of observed vs predicted values,
reporting the equation and R2 value.
}
