% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_C.R
\name{pred_C}
\alias{pred_C}
\title{A function to predict excretion rate according to Wilson et al 2009 (p = 1 and p = 2.4) and our model}
\usage{
pred_C(w, t, ril, ar)
}
\arguments{
\item{w}{Body mass (g)}

\item{t}{Temperature (°C)}

\item{ril}{Relative intestinal length}

\item{ar}{Caudal fin aspect ratio}
}
\description{
A function to predict excretion rate according to Wilson et al 2009 (p = 1 and p = 2.4) and our model
}
