% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q10.R
\name{Q10}
\alias{Q10}
\title{Compute Q10 temperature coefficient}
\usage{
Q10(r1, r2, t1, t2)
}
\arguments{
\item{r1}{rate at temperature 1}

\item{r2}{rate at temperature 2}

\item{t1}{temperature 1}

\item{t2}{temperature 2}
}
\description{
Compute Q10 temperature coefficient
}
\examples{
\dontrun{
Q10(r1 = 2.2, r2 = 3.6, t1 = 15, t2 = 24)}

}
