% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_caco3_models.R
\name{run_caco3_models}
\alias{run_caco3_models}
\title{Run a set of 20 models on species-level carbonate excretion rates}
\usage{
run_caco3_models(data, ...)
}
\arguments{
\item{data}{A data frame containing 9 columns:
"log_exc_rate": natural-log transformed carbonate excretion rate in "µmol/h";
"log_weight": natural-log transformed fish weight in "kg";
"log_ril": natural-log transformed relative intestinal length;
"sqrt_asp_ratio": square-root transformed aspect ratio of caudal fin;
"mean_S": salinity;
"mean_T": water temperature in "°C";
"family": taxonomic family;
"sqrt_sampling": square-root transformed sampling interval in hours
"weights": the weight of each observation.}

\item{...}{Additional parameters passed to \code{\link[brms:brm]{brms::brm()}}}
}
\value{
A tibble with four columns:
"id: the model names;
"bf": the model formulas;
"prior": the priors;
"fit": the model objects.
}
\description{
Run a set of 20 models on species-level carbonate excretion rates
}
