% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_intestinal_length.R
\name{test_intestinal_length}
\alias{test_intestinal_length}
\title{Test extrapolation of intestinal length}
\usage{
test_intestinal_length(
  data,
  method = "fitted",
  ndraws = 1000,
  seed = 1,
  mc.cores = getOption("mc.cores", 1L),
  summary = TRUE,
  .point = mean,
  .width = 0.95
)
}
\arguments{
\item{data}{A data frame with five columns:
\itemize{
\item{"id":} an identifier for each observation
\item{"species":} the taxonomic name in the format "Genus species"
\item{"sl":} standard length in "mm"
\item{"elongation":} body elongation (i.e. standard length/max body depth)
\item{"trophic_level":} trophic level (from 1 to 5, preferably based on food items)
}}

\item{method}{Either 'fitted' or 'predict'. Predictions performed with 'predict'
have higher variance as they incorporate the residual error.
However, the estimated means of both methods should be very similar}

\item{ndraws}{The number of draws to be used for the predictions}

\item{seed}{The seed for random number generation to make results reproducible.
Must be different from NULL}

\item{mc.cores}{The number of cores to use for
\code{\link[parallel:mclapply]{parallel::mclapply()}}.
Must be exactly 1 on Windows (which uses the master process)}

\item{summary}{Logical, if predictions should be summarised.
If TRUE returns mean and sd, if FALSE returns all draws}

\item{.point}{Only if 'summary=TRUE', point summary function,
e.g. 'mean' or 'median', default to 'mean'}

\item{.width}{Only if 'summary=TRUE', according to
\code{\link[tidybayes:point_interval]{tidybayes::point_interval()}}}
}
\value{
A data frame with natural-log intestinal length in "mm"
}
\description{
Test extrapolation of intestinal length
}
