% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_intestinal_length.R
\name{test_phylo_effect}
\alias{test_phylo_effect}
\title{Test estimation of phylogenetic effect for intestinal length
using observed or unobserved taxa}
\usage{
test_phylo_effect(
  ndraws = NULL,
  seed = NULL,
  mc.cores = getOption("mc.cores", 1L),
  phy,
  unsampled = NULL
)
}
\arguments{
\item{ndraws}{The number of draws to return, or NULL to return all draws}

\item{seed}{A seed to use when subsampling draws (i.e. when ndraw is not NULL)}

\item{mc.cores}{The number of cores to use for \code{\link[parallel:mclapply]{parallel::mclapply()}}.
Must be exactly 1 on Windows (which uses the master process)}

\item{phy}{An object of class "phylo" or "multiPhylo" including the new taxa
for which the phylogenetic effect has to be predicted}

\item{unsampled}{Vector of species (observed or unobserved) for
which the phylogenetic effect has to be predicted}
}
\value{
A data frame
}
\description{
Test estimation of phylogenetic effect for intestinal length
using observed or unobserved taxa
}
