% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trophic_level_FB.R
\name{trophic_level_FB}
\alias{trophic_level_FB}
\title{Obtain trophic level from FishBase}
\usage{
trophic_level_FB(
  species_list,
  type = c("both", "diet", "food items"),
  check_names = FALSE,
  mc.cores = getOption("mc.cores", 1L)
)
}
\arguments{
\item{species_list}{A vector of taxonomic names in the form "Genus species"}

\item{type}{Character, indicating the type of estimate desired. Possible values are:
"both" (default), "diet", "food items" or any abbreviation of these (see Details)}

\item{check_names}{Logical, check whether species names are correct
according to the accepted taxonomy in FishBase}

\item{mc.cores}{The number of cores to use for \code{\link[parallel:mclapply]{parallel::mclapply()}}.
Must be exactly 1 on Windows (which uses the master process)}
}
\value{
A tibble with species name, trophic level, and taxonomic level.
}
\description{
Obtain trophic level from FishBase at the lowest available taxonomic level,
either species, genus or family level.
}
\details{
The function can retrieve three different trophic level estimates:
\itemize{
\item{\code{type="diet"}:} retrieves trophic level based on diet composition only. These estimates are available only for some species. Note that for some families there are no diet studies available;
\item{\code{type="food items"}:} retrieves trophic level based on food items only;
\item{\code{type="both"}:} computes the average of the two when both are available, otherwise return the available estimate.
}
}
\examples{
\dontrun{
trophic_level_FB("Cephalopholis argus", type = "diet")
trophic_level_FB(c("Acanthurus achilles", "Epibulus insidiator"), type = "food items")
library(rfishbase)
scombrids <- species_list(Family = "Scombridae")
trophic_level_FB(scombrids, type = "both")
}

}
