% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trophic_level_FB.R
\name{trophic_level_FB_sp}
\alias{trophic_level_FB_sp}
\title{Obtain trophic level from FishBase for taxa identified to genus level}
\usage{
trophic_level_FB_sp(
  species_list,
  type = c("both", "diet", "food items"),
  mc.cores = getOption("mc.cores", 1L)
)
}
\arguments{
\item{species_list}{A vector of taxonomic names in the form "Genus sp." or "Genus spp."}

\item{type}{Character, indicating the type of estimate desired. Possible values are:
"both" (default), "diet", "food items" or any abbreviation of these (see Details)}

\item{mc.cores}{The number of cores to use for \code{\link[parallel:mclapply]{parallel::mclapply()}}.
Must be exactly 1 on Windows (which uses the master process)}
}
\value{
A tibble with species name, trophic level, and taxonomic level.
}
\description{
Obtain trophic level from FishBase at the lowest available taxonomic level,
either genus or family level.
}
\details{
The function can retrieve three different trophic level estimates:
\code{type="diet"} retrieves trophic level based on diet composition only. These estimates are available only for some species. Note that for some families there are no diet studies available;
\code{type="food items"} retrieves trophic level based on food items only;
\code{type="both"} computes the average of the two when both are available, otherwise return the available estimate.
}
\examples{
\dontrun{
trophic_level_FB_sp("Cephalopholis sp.", type = "diet")
trophic_level_FB_sp(c("Acanthurus sp.", "Lethrinus spp."), type = "food items")
}

}
